use module logic
