+++ src/partials/home_page/home_page.html
@@ -38,7 +38,7 @@
     >
   </div>
   <div class="gallery">
-    <div class="newsHomePage-card">
+    <!-- <div class="newsHomePage-card">
       <div class="card-picture">
         <img
           class="newsHomePage-image"
@@ -163,7 +163,7 @@
         <p class="newsHomePage-date">24/02/2023</p>
         <a class="newsHomePage-readMore-link" href="">Read more</a>
       </div>
-    </div>
+    </div> -->
   </div>
   <div class="pagination-buttons">
     <button class="prew-btn" type="button">Prew</button>
